requirements.txt
# requirements.txt
streamlit>=1.25.0
pandas>=2.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
Pillow>=9.5.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0